server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
beanie==2.0.0
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.3
h11==0.16.0
idna==3.11
lazy-model==0.3.0
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.38.0
scikit-learn
pandas
joblib
argon2-cffi
httpx
